+++ public/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v4.0.28-beta (17.03.2024)
+## Panel
+- bugfix for hide/show Levels
+- bugfix for maximizable button
+## Documentation
+- Better explanation of the syntax for definitions
 # v4.0.27-beta (16.03.2024)
 ## Core
 - replacer %n option noLocaleNumber deprecated -> removed!